4_agentic_labs/openclaw_network_boundary/services/mitmproxy/requirements.txt
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
